requirements.txt
appnope==0.1.4
asgiref==3.7.2
astroid==3.0.2
asttokens==2.4.1
autoflake==2.3.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
debugpy==1.8.9
decorator==5.1.1
dill==0.3.7
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djeasyview==1.0.5
drf-yasg==1.21.8
executing==2.1.0
ghp-import==2.1.0
gunicorn==21.2.0
idna==3.10
inflection==0.5.1
ipykernel==6.29.5
ipython==8.29.0
isort==5.13.2
jedi==0.19.2
Jinja2==3.1.3
jupyter_client==8.6.3
jupyter_core==5.7.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mccabe==0.7.0
mergedeep==1.3.4
mkdocs==1.5.3
nest-asyncio==1.6.0
packaging==23.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.1.0
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.10.0
pylint==3.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyyaml_env_tag==0.1
pyzmq==26.2.0
requests==2.32.3
sentry-sdk==1.39.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomlkit==0.12.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.1.0
watchdog==4.0.0
wcwidth==0.2.13
